sql: split sql-agnostic, and postgresql methods.
